better cli argument parsing
